--- a/wire/gen.go
+++ b/wire/gen.go
@@ -1,0 +1,3 @@
+package wire
+
+//go:generate protoc --gofast_out=$GOPATH/src/github.com/opentracing/basictracer-go/wire wire.proto