--- a/wire/wire.proto
+++ b/wire/wire.proto
@@ -1,0 +1,13 @@
+syntax = "proto3";
+package basictracer_go.wire;
+option go_package = "wire";
+
+message TracerState {
+  int64 trace_id = 1;
+  int64 span_id = 2;
+  bool  sampled = 3;
+}
+
+message Baggage {
+  map<string, string> items = 1;
+}